# openapi.json
{"openapi":"3.1.0","info":{"title":"ETRIS Ultimate AI Demo API","version":"0.1.0"},"paths":{"/api/anpr/status":{"get":{"tags":["anpr"],"summary":"Status","operationId":"status_api_anpr_status_get","parameters":[{"name":"mode","in":"query","required":false,"schema":{"type":"string","default":"LIVE_ANALYSIS","title":"Mode"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/anpr/events":{"get":{"tags":["anpr"],"summary":"Events","operationId":"events_api_anpr_events_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/anpr/recognition":{"get":{"tags":["anpr"],"summary":"Recognition","operationId":"recognition_api_anpr_recognition_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/anpr/stream":{"get":{"tags":["anpr"],"summary":"Stream","operationId":"stream_api_anpr_stream_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/anpr/recorded-video":{"get":{"tags":["anpr"],"summary":"Recorded Video","operationId":"recorded_video_api_anpr_recorded_video_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/anpr/recorded-events":{"get":{"tags":["anpr"],"summary":"Recorded Events","operationId":"recorded_events_api_anpr_recorded_events_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/anpr/play":{"post":{"tags":["anpr"],"summary":"Play","operationId":"play_api_anpr_play_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/anpr/pause":{"post":{"tags":["anpr"],"summary":"Pause","operationId":"pause_api_anpr_pause_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/anpr/restart":{"post":{"tags":["anpr"],"summary":"Restart","operationId":"restart_api_anpr_restart_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/anpr/processing-rate":{"post":{"tags":["anpr"],"summary":"Processing Rate","operationId":"processing_rate_api_anpr_processing_rate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProcessingRateRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/sightings/recent":{"get":{"tags":["stage9"],"summary":"Recent","operationId":"recent_api_sightings_recent_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":500,"minimum":1,"default":50,"title":"Limit"}},{"name":"camera_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Camera Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/sightings/{plate_text}":{"get":{"tags":["stage9"],"summary":"By Plate","operationId":"by_plate_api_sightings__plate_text__get","parameters":[{"name":"plate_text","in":"path","required":true,"schema":{"type":"string","title":"Plate Text"}},{"name":"start_time","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Start Time"}},{"name":"end_time","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"End Time"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/cameras":{"get":{"tags":["stage9"],"summary":"Cameras","operationId":"cameras_api_cameras_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/trajectory/{plate_text}":{"get":{"tags":["stage9"],"summary":"Trajectory","operationId":"trajectory_api_trajectory__plate_text__get","parameters":[{"name":"plate_text","in":"path","required":true,"schema":{"type":"string","title":"Plate Text"}},{"name":"start_time","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Start Time"}},{"name":"end_time","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"End Time"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/analytics/summary":{"get":{"tags":["analytics"],"summary":"Summary","operationId":"summary_api_analytics_summary_get","parameters":[{"name":"start_time","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Start Time"}},{"name":"end_time","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"End Time"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/analytics/camera-volume":{"get":{"tags":["analytics"],"summary":"Camera Volume","operationId":"camera_volume_api_analytics_camera_volume_get","parameters":[{"name":"start_time","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Start Time"}},{"name":"end_time","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"End Time"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/analytics/od":{"get":{"tags":["analytics"],"summary":"Od","operationId":"od_api_analytics_od_get","parameters":[{"name":"start_time","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Start Time"}},{"name":"end_time","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"End Time"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/analytics/segments":{"get":{"tags":["analytics"],"summary":"Segments","operationId":"segments_api_analytics_segments_get","parameters":[{"name":"start_time","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Start Time"}},{"name":"end_time","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"End Time"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/analytics/congestion":{"get":{"tags":["analytics"],"summary":"Congestion","operationId":"congestion_api_analytics_congestion_get","parameters":[{"name":"start_time","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Start Time"}},{"name":"end_time","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"End Time"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/analytics/heatmap":{"get":{"tags":["analytics"],"summary":"Heatmap","operationId":"heatmap_api_analytics_heatmap_get","parameters":[{"name":"start_time","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Start Time"}},{"name":"end_time","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"End Time"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/analytics/routes":{"get":{"tags":["analytics"],"summary":"Routes","operationId":"routes_api_analytics_routes_get","parameters":[{"name":"start_time","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Start Time"}},{"name":"end_time","in":"query","required":false,"schema":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"End Time"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/signal-control/demo":{"get":{"tags":["signal-control"],"summary":"Demo","operationId":"demo_api_signal_control_demo_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/signal-control/demo/change":{"get":{"tags":["signal-control"],"summary":"Demo Change","operationId":"demo_change_api_signal_control_demo_change_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/signal-control/plan":{"post":{"tags":["signal-control"],"summary":"Plan","operationId":"plan_api_signal_control_plan_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlanPayload"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/signal-control/traffic-demo/events":{"get":{"tags":["signal-control"],"summary":"Traffic Demo Events","operationId":"traffic_demo_events_api_signal_control_traffic_demo_events_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/signal-control/traffic-demo/video":{"get":{"tags":["signal-control"],"summary":"Traffic Demo Video","operationId":"traffic_demo_video_api_signal_control_traffic_demo_video_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/alerts":{"get":{"tags":["alerts"],"summary":"Alerts","operationId":"alerts_api_alerts_get","parameters":[{"name":"alert_type","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Alert Type"}},{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"}},{"name":"camera_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Camera Id"}},{"name":"zone_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Zone Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/alerts/active":{"get":{"tags":["alerts"],"summary":"Active","operationId":"active_api_alerts_active_get","parameters":[{"name":"alert_type","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Alert Type"}},{"name":"camera_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Camera Id"}},{"name":"zone_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Zone Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/alerts/{alert_id}":{"get":{"tags":["alerts"],"summary":"By Id","operationId":"by_id_api_alerts__alert_id__get","parameters":[{"name":"alert_id","in":"path","required":true,"schema":{"type":"string","title":"Alert Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ApproachPayload":{"properties":{"approach_id":{"type":"string","title":"Approach Id"},"vehicle_count":{"type":"number","title":"Vehicle Count"},"queue_length":{"type":"number","title":"Queue Length"},"lane_occupancy":{"type":"number","title":"Lane Occupancy"},"average_waiting_time_s":{"type":"number","title":"Average Waiting Time S"},"arrival_rate_vpm":{"type":"number","title":"Arrival Rate Vpm"},"heavy_vehicle_count":{"type":"number","title":"Heavy Vehicle Count"},"timestamp":{"type":"string","format":"date-time","title":"Timestamp"},"consecutive_cycles_skipped":{"type":"integer","title":"Consecutive Cycles Skipped","default":0},"time_since_last_green_s":{"type":"number","title":"Time Since Last Green S","default":0}},"type":"object","required":["approach_id","vehicle_count","queue_length","lane_occupancy","average_waiting_time_s","arrival_rate_vpm","heavy_vehicle_count","timestamp"],"title":"ApproachPayload"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"PlanPayload":{"properties":{"intersection_id":{"type":"string","title":"Intersection Id"},"approaches":{"items":{"$ref":"#/components/schemas/ApproachPayload"},"type":"array","title":"Approaches"}},"type":"object","required":["intersection_id"],"title":"PlanPayload"},"ProcessingRateRequest":{"properties":{"rate":{"type":"string","title":"Rate"}},"type":"object","required":["rate"],"title":"ProcessingRateRequest"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}